feat: GPU bench round — tree verify, distilled draft, int8/int4 KV decode kernels
- bench_tree_verify: tree verify 2.1x loses to linear lookup 2.4x (accept
  rate falls with candidate width); tree token-identical, lookup divergence
  diagnosed as fp16 near-tie (margin 0.0156)
- bench_distilled_draft: LoRA logit-KD lifts accept 49.8% -> 52.4%; whole
  speculative path still loses to greedy on 3B/one-GPU (honest negative)
- kernels: fused attention_decode_quant (int8 + packed int4, split-K),
  ~1.6-2.5x over fp16 KV at T=262k, not roofline 4x (int4 unpack
  instruction-bound; <64k launch/merge floor) + tests
- logit_kd: kd_loss upcasts to fp32 (fp16 log_softmax NaN'd Adam on a real
  fp16 teacher)
- eval_ruler: RULER runner w/ chunked prefill (one-shot 20k prefill OOMs
  on 10GB); Qwen2.5-3B 100% on all tasks through 20k tokens

+++ llmopt/distill/logit_kd.py
@@ -22,10 +22,12 @@ def kd_loss(student_logits, teacher_logits, temperature: float = 1.0):
     import torch.nn.functional as F
 
     t = temperature
+    # softmax/KL in fp32: fp16 log_softmax under-/overflows enough to NaN
+    # Adam on real models (seen with a fp16 Qwen teacher)
     return (
         F.kl_div(
-            F.log_softmax(student_logits / t, dim=-1),
-            F.log_softmax(teacher_logits / t, dim=-1),
+            F.log_softmax(student_logits.float() / t, dim=-1),
+            F.log_softmax(teacher_logits.float() / t, dim=-1),
             log_target=True,
             reduction="batchmean",
         )
